requirements.txt
python-fasthtml
fastlite
httpx
langgraph
langchain
langchain-openai
langchain-community
openai
yfinance
pandas
numpy
requests
python-dotenv
psycopg2-binary
bcrypt
authlib
pytest
pytest-asyncio
pydantic
typing-extensions
plotly
networkx
python-pptx
openpyxl
pdfplumber
exa-py
tavily-python
